+++ src/pages/schedule/components/ListCard.tsx
@@ -1,27 +1,45 @@
 import CommonCard from '@/components/common/CommonCard.tsx'
 import DayBadgeWithTitle from '@/components/common/DayBadgeWithTitle.tsx'
 import { NextIcon } from '@/icons'
+import useDateInfo from '@/hooks/useDateInfo.ts'
+import { WeatherGroup, WeatherProps } from '@/components/common/Weather.tsx'
+import { useNavigate } from 'react-router-dom'
 
-const ListCard = () => {
+const weathers: WeatherProps[] = [
+  { weather: 'blizzard', isToday: false, date: '2024-07-27T15:24:00', temperature: 30 },
+  { weather: 'blizzard', isToday: false, date: '2024-07-27T15:24:00', temperature: 30 },
+  { weather: 'blizzard', isToday: false, date: '2024-07-27T15:24:00', temperature: 30 },
+  { weather: 'blizzard', isToday: false, date: '2024-07-27T15:24:00', temperature: 30 },
+  { weather: 'sunny', isToday: true, date: '2024-07-27T15:24:00', temperature: 30 },
+  { weather: 'blizzard', isToday: false, date: '2024-07-27T15:24:00', temperature: 30 },
+]
+const ListCard = ({ schedule }) => {
+  const navigate = useNavigate()
+  const { formattedDate, dDayText } = useDateInfo(schedule?.scheduleDate)
   return (
     <CommonCard>
       <div className="flex flex-col gap-1">
         <DayBadgeWithTitle
-          text="D-day "
-          title="5월 26일 (수)"
+          text={dDayText}
+          title={formattedDate}
           rightAction={
-            <button className="flex items-center gap-[7px] text-b2">
+            <button
+              onClick={() => navigate(`/schedule/detail/${schedule?.scheduleId}`)}
+              className="flex items-center gap-[7px] text-b2"
+            >
               자세히
               <NextIcon />
             </button>
           }
         />
         <div className="flex gap-1">
-          <div className="text-b2 font-semibold">북한산</div>
-          <div className="text-b2">백운대코스</div>
+          <div className="text-b2 font-semibold">{schedule?.mountain}</div>
+          <div className="text-b2">{schedule?.course}</div>
         </div>
       </div>
-      <div className="pt-5">날씨</div>
+      <div className="pt-5">
+        <WeatherGroup weathers={weathers} className="mt-[10px]" />
+      </div>
     </CommonCard>
   )
 }
