+++ src/pages/schedule/list/index.tsx
@@ -1,11 +1,10 @@
 import Header from '@/components/layouts/header'
-import { useNavigate } from 'react-router-dom'
 import ListCard from '@/pages/schedule/components/ListCard.tsx'
 import { useQuery } from '@tanstack/react-query'
 import { getScheduleList } from '@/services/api/schedule'
+import SuggestionPrompt from '@/pages/schedule/list/components/SuggestionPrompt.tsx'
 
 const ListSchedule = () => {
-  const navigate = useNavigate()
   const { data, isError } = useQuery({
     queryKey: ['scheduleList'],
     queryFn: getScheduleList,
@@ -14,22 +13,12 @@ const ListSchedule = () => {
   return (
     <div className="flex h-full flex-col">
       <Header title="등산 일정" rightAction={<button className="text-b2 text-primary">추가</button>} />
-      <div className="flex h-full flex-col justify-between bg-background p-5">
-        <ListCard />
-        <div className="flex h-full flex-col items-center justify-center">
-          <img alt="등산 일정 이미지" src="src/assets/images/mountainListImg.png" />
-          <div className="mb-2 mt-5 text-h5 text-green-950">일정을 정하고 친구와 공유해요</div>
-          <div className="mb-4 text-center text-b2 text-gray-600">
-            등산하기 좋은 산을 알려드릴게요
-            <br />산 구경하고 일정 정해볼까요 ?
-          </div>
-          <button
-            className="rounded-[32px] border-[1px] border-border bg-white px-3 py-2 text-b2"
-            onClick={() => navigate('/home')}
-          >
-            구경하러 가기
-          </button>
-        </div>
+      <div className="flex h-full flex-col gap-3 bg-background p-5">
+        {data && data.length > 0 ? (
+          data.map((schedule, index) => <ListCard key={index} schedule={schedule} />)
+        ) : (
+          <SuggestionPrompt />
+        )}
       </div>
     </div>
   )
